requirements.txt
rio-ui
psycopg2-binary
sqlalchemy
python-dotenv
bcrypt
python-jose[cryptography]
